refactor: split support reference ingestion dtos
Evidence:

- python -m pytest tests/unit/services/ingestion_service/test_reference_data_dto.py -q: 35 passed

- python -m ruff check scoped DTO/test files: passed

- python -m ruff format --check scoped DTO/test files: passed

- python -m mypy src/services/ingestion_service/app/DTOs/reference_data_support_dto.py: passed

- radon mi: reference_data_dto.py improved from B (12.49) to B (14.29); reference_data_support_dto.py A (43.83)
